twillio/requirements.txt
fastapi
uvicorn[standard]
requests
twilio
python-dotenv
python-multipart
